api/requirements.txt
azure-functions
azure-cosmos
azure-storage-blob
requests
